requirements.txt
pystray 
pillow 
pywin32
customtkinter